+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v1.2.0
+
++ Integrated [Unidragger](https://github.com/metafizzy/unidragger)
+  - Changed `dragStart` to trigger when dragging starts, rather than on pointer down. Fixed [#13](https://github.com/desandro/draggabilly/issues/13).
+  - Add `staticClick` event. Fixed [#19](https://github.com/desandro/draggabilly/issues/13).
++ Added jQuery plugin support. Fixed [#73](https://github.com/desandro/draggabilly/issues/73).
++ Removed `draggieInstance` first argument from event callbacks
+
 #### v1.1.2
 
 Add `destroy` method. Fixes [#72](https://github.com/desandro/draggabilly/issues/72)